+++ doc/CHANGELOG.rst
@@ -21,6 +21,9 @@ Latest Changes:
 
   - Fixed a bug in which JPype did not respect a JConversion between two Java classes.
 
+  - Enhancement in convertToDirectBuffer to support wrapping bytes and readonly
+    memoryviews as readonly java ByteBuffers.
+
 - **1.5.2 - 2025-01-20**
 
   - Roll back agent change due to misbehaving JVM installs.